agcms-auth/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncpg==0.29.0
redis[asyncio]==5.0.3
